# Conflicts:
#	Ecommerce.API/Ecommerce.API.csproj
#	Ecommerce.API/Program.cs
#	Ecommerce.API/appsettings.json
#	Ecommerce.Data/Contexts/EcommerceContext.cs
#	Ecommerce.Data/Ecommerce.Data.csproj
#	Ecommerce.Data/Migrations/20250125230904_FirstMigration.Designer.cs
#	Ecommerce.Data/Migrations/20250125230904_FirstMigration.cs
#	Ecommerce.Data/Migrations/EcommerceContextModelSnapshot.cs
#	Ecommerce.Models/Ecommerce.Models.csproj
#	Ecommerce.Models/Usuario.cs

+++ Ecommerce.API/Controllers/UsuarioController.cs
@@ -3,6 +3,8 @@ using Ecommerce.Models;
 using Ecommerce.Models.Dtos;
 using Microsoft.AspNetCore.Mvc;
 using Ecommerce.Models.Dtos.DtoMappers;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Cors;
 
 
 namespace Ecommerce.API.Controllers
@@ -26,27 +28,43 @@ namespace Ecommerce.API.Controllers
         {
             try
             {
-                var usuarioCreado = await _usuarioService.AddUsuario(usuario.UsuarioDtoAUsuario());
-                return Ok(usuarioCreado);
+                //Intenta crear el usuario
+                var usuarioCreado = await _usuarioService.AddUsuario(usuario);
+
+                //Si todo sale bien, devuelve un ok con el usuario creado recientemente
+                return Ok(usuarioCreado.usuarioADto());
             }
-            catch
-            {
-                return BadRequest("No se pudo agregar el usuario");
+            catch (DbUpdateException ex)
+            {   
+                //Captura errores al guardar los datos en la DB
+                return BadRequest(new { Error = "Error al guardar los datos en la base de datos", Details = ex.InnerException?.Message ?? ex.Message });
+            }
+            catch (Exception ex)
+            {   
+                //Captura otro tipo de errores
+                return BadRequest(new { Error = "Ocurrió un error al procesar la solicitud", Details = ex.Message });
             }
         }
 
-        [HttpGet("get-all")]
 
+        //Metodo para mostrar toda la lista de usuarios de la Base de Datos
+        [HttpGet("get-all")]
         public async Task<IActionResult> GetAll() 
         {
 
             try
             {
-
+                //Crea una lista de usuarios por medio de la funcion GetUsuarios del service
                 var listaUsuarios = await _usuarioService.GetUsuarios();
-                return Ok(listaUsuarios);
+
+                //Creada la lista la modifica de usuario a usuarioDto 
+                var listaUsuariosDto = listaUsuarios.Select(u => u.usuarioADto()).ToList();
+
+                //Retorna la lista de usuarioDtos
+                return Ok(listaUsuariosDto);
 
             }
+            //En caso de que ocurra un error lo captura y lo muestra
             catch (Exception ex) 
             {
 
@@ -57,6 +75,7 @@ namespace Ecommerce.API.Controllers
         
         }
 
+        //Metodo para mostrar un usuario especifico
         [HttpGet("{id}")]
         public async Task<IActionResult> GetUsuarioById(Guid id)
         {
@@ -68,55 +87,72 @@ namespace Ecommerce.API.Controllers
                 // Verifica si el usuario es nulo
                 if (usuario == null)
                 {
+                    //Si el usuario no se encuentra muestra un error y el siguiente mensaje
                     return NotFound(new { Message = $"El id '{id}' no corresponde a ningún usuario en la base de datos." });
                 }
 
-               
-                return Ok(usuario);
+               //Si puede encontrar el usuario sin problemas lo retorna en formato dto
+                return Ok(usuario.usuarioADto());
             }
             catch (Exception ex)
             {
-               
+               //Captura errores internos de la Base de Datos
                 return StatusCode(500, new { Error = "Ocurrió un error interno.", Details = ex.Message });
             }
         }
 
-        //Siempre sale por el catch
+
+        //Metodo para eliminar un suario especifico
         [HttpDelete("id")]
         public async Task<IActionResult> DeleteUser(Guid id) 
         {
             try
             {
+                //Crea una variable para poder asignar el usuario en caso de que exista en la base de datos
                 var usuario = new Usuario();
                 usuario = await _usuarioService.GetUsuario(id);
+
+
+                if(usuario == null)
+                    //Si no encuentra el usuario a borrar muestra el notfound y el mensaje de error
+                    return NotFound(new { Message = $"El id '{id}' no corresponde a ningún usuario en la base de datos." });
+
+                
+                //Si el usuario se encuentra en la base de datos lo borra y retorna el usuario que se acaba de eliminar en formato dto
                 await _usuarioService.DeleteUsuario(id);                
-                return Ok(usuario);
+                return Ok(usuario.usuarioADto());
 
             }
-            catch  
+            catch (Exception ex)
             {
-
-                return NotFound(new { Message = $"El id '{id}' no corresponde a ningún usuario en la base de datos." });
-            
+                //Captura errores internos de la Base de Datos
+                return StatusCode(500, new { Error = "Ocurrió un error interno.", Details = ex.Message });
             }
-        
+
         }
+
+
+        //Metodo para modificar usuarios
         [HttpPut("id")]
         public async Task<IActionResult> ModificarUsuario(Guid id, UsuarioDto usuario)
         {
             try
             {
+                //Intenta hacer la modificacion al usuario
                 var resultado = await _usuarioService.UpdateUsuario(id ,usuario.UsuarioDtoAUsuario());
+                //Si la modificacion fue exitosa devuelve verdadero 
+               
                 if (!resultado)
                 {
+                    //Si no se pudo realizar la modificacion retorna el error
                     return NotFound(new { Message = $"El id '{id}' no corresponde a ningún usuario en la base de datos." });
                 }
-                return Ok(usuario);
+
+                //Retorno del ok y el usuario modificado
+                return Ok("Usuario modificado con exito");
             }
-            catch (KeyNotFoundException ex)
-            {
-                return NotFound(new { Message = ex.Message });
-            }
+
+            //Captura errores y los retorna
             catch (Exception ex)
             {
                 return StatusCode(500, new { Message = "Error interno del servidor", Details = ex.Message });
